Add function to get hypervisor-specific builderFactory
