+++ cmd/virt-launcher/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "//pkg/ephemeral-disk:go_default_library",
         "//pkg/hooks:go_default_library",
         "//pkg/hotplug-disk:go_default_library",
+        "//pkg/hypervisor:go_default_library",
         "//pkg/ignition:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/virt-config:go_default_library",
