+++ pkg/hypervisor/BUILD.bazel
@@ -15,9 +15,6 @@ go_library(
         "//pkg/virt-handler/isolation:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter:go_default_library",
-        "//pkg/virt-launcher/virtwrap/converter/compute:go_default_library",
-        "//pkg/virt-launcher/virtwrap/converter/metadata:go_default_library",
-        "//pkg/virt-launcher/virtwrap/converter/network:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
