+++ pkg/virt-launcher/virtwrap/live-migration-target.go
@@ -42,7 +42,6 @@ import (
 	diskutils "kubevirt.io/kubevirt/pkg/ephemeral-disk-utils"
 	cmdv1 "kubevirt.io/kubevirt/pkg/handler-launcher-com/cmd/v1"
 	"kubevirt.io/kubevirt/pkg/hooks"
-	"kubevirt.io/kubevirt/pkg/hypervisor"
 	"kubevirt.io/kubevirt/pkg/pointer"
 	"kubevirt.io/kubevirt/pkg/util"
 	"kubevirt.io/kubevirt/pkg/util/net/ip"
@@ -173,9 +172,8 @@ func (l *LibvirtDomainManager) prepareMigrationTarget(
 		return fmt.Errorf("Failed to generate libvirt domain from VMI spec: %v", err)
 	}
 
-	domainBuilderFactory := hypervisor.KvmDomainBuilderFactory{} // TODO Get this from virt-launcher cmdline
 	domain := &api.Domain{}
-	if err := converter.Convert_v1_VirtualMachineInstance_To_api_Domain(vmi, domain, domainBuilderFactory.MakeDomainBuilder(c), c); err != nil {
+	if err := converter.Convert_v1_VirtualMachineInstance_To_api_Domain(vmi, domain, l.domainBuilderFactory.MakeDomainBuilder(c), c); err != nil {
 		return fmt.Errorf("conversion failed: %v", err)
 	}
