+++ pkg/virt-launcher/virtwrap/util/BUILD.bazel
@@ -36,6 +36,7 @@ go_test(
     deps = [
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
         "//pkg/hooks:go_default_library",
+        "//pkg/hypervisor:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
